en teoria todo igual pero modularizado
