+++ frontend/src/components/atoms/PolygonDrawerControl.tsx
@@ -1,0 +1,24 @@
+import { ControlButtons } from '../ui/button'
+
+interface PolygonDrawerControlProps {
+  isDrawing: boolean
+  polygonPoints: [number, number][]
+  onToggleDrawing: () => void
+  onSearch: () => void
+}
+
+export function PolygonDrawerControl({
+  isDrawing,
+  polygonPoints,
+  onToggleDrawing,
+  onSearch,
+}: PolygonDrawerControlProps) {
+  return (
+    <ControlButtons
+      isDrawing={isDrawing}
+      polygonPoints={polygonPoints}
+      onToggleDrawing={onToggleDrawing}
+      onSearch={onSearch}
+    />
+  )
+}
